refactor: split user-management.ts to comply with 200-line limit
